API Move elemental editor actions menu items into their own HOCs
