+++ client/src/boot/registerTransforms.js
@@ -3,6 +3,9 @@ import readOneBlockQuery from 'state/history/readOneBlockQuery';
 import HistoricElementViewFactory from 'components/HistoricElementView/HistoricElementView';
 import revertToBlockVersionMutation from 'state/history/revertToBlockVersionMutation';
 import readBlocksForPageQuery from 'state/editor/readBlocksForPageQuery';
+import withArchiveAction from 'components/ElementActions/ArchiveAction';
+import withPublishAction from 'components/ElementActions/PublishAction';
+import withUnpublishAction from 'components/ElementActions/UnpublishAction';
 
 export default () => {
   Injector.transform(
@@ -54,4 +57,11 @@ export default () => {
       );
     }
   );
+
+  // Add elemental editor actions
+  Injector.transform('element-actions', (updater) => {
+    updater.component('ElementActions', withArchiveAction, 'ElementActionsWithArchive');
+    updater.component('ElementActions', withPublishAction, 'ElementActionsWithPublish');
+    updater.component('ElementActions', withUnpublishAction, 'ElementActionsWithUnpublish');
+  });
 };
